dj_checkmate/requirements.txt
wheel
typing

django
djangorestframework
django-filter
django-cors-headers